// src/types/product.ts
// const API 

export type Product = {
    name: string 
    description: string
    price: number
    imageData: string
}

export type ProductRecord = {
    id: string 
    name: string 
    description: string 
    price: number 
    imageUrl: string 
    createdAt: string 
    updatedAt: string 
}